requirements.txt
torch
torchvision
timm
faiss-cpu
streamlit
numpy
Pillow
